feat: 修复CypherQuery mock 数据
